refactor: move demo out of the migrator package
